moar refactoring
